requirements.txt
aiogram>=3.7.0,<3.8
APScheduler>=3.10.4,<3.11
matplotlib>=3.8.0,<3.9
python-dotenv>=1.0.1,<2.0
